refactor file and stats into utils
